all: Remove Echo
