+++ go.mod
@@ -17,10 +17,6 @@ replace google.golang.org/grpc => google.golang.org/grpc v1.33.1
 // Do not upgrade genproto beyond v0.0.0-20200513103714-09dca8ec2884
 replace google.golang.org/genproto => google.golang.org/genproto v0.0.0-20200513103714-09dca8ec2884
 
-// Do not upgrade Echo beyond v4.1.2.
-// See https://github.com/TheThingsNetwork/lorawan-stack/issues/977.
-replace github.com/labstack/echo/v4 => github.com/labstack/echo/v4 v4.1.2
-
 // Do not upgrade go-sqlmock beyond v1.3.0.
 // See https://github.com/heptiolabs/healthcheck/issues/23.
 replace gopkg.in/DATA-DOG/go-sqlmock.v1 => gopkg.in/DATA-DOG/go-sqlmock.v1 v1.3.0
@@ -44,8 +40,6 @@ replace github.com/go-redis/redis/v8 => github.com/go-redis/redis/v8 v8.4.0
 
 require (
 	contrib.go.opencensus.io/exporter/prometheus v0.1.0
-	github.com/PuerkitoBio/purell v1.1.1
-	github.com/PuerkitoBio/urlesc v0.0.0-20170810143723-de5bf2ad4578 // indirect
 	github.com/TheThingsIndustries/mystique v0.0.0-20200127144137-4aa959111fe7
 	github.com/TheThingsIndustries/protoc-gen-go-json v1.1.3
 	github.com/TheThingsNetwork/go-cayenne-lib v1.1.0
@@ -64,7 +58,6 @@ require (
 	github.com/getsentry/sentry-go v0.11.0
 	github.com/go-redis/redis/v8 v8.4.0
 	github.com/gogo/protobuf v1.3.2
-	github.com/golang/gddo v0.0.0-20210115222349-20d68f94ee1f
 	// NOTE: github.com/golang/protobuf is actually pinned to v1.3.5 above.
 	github.com/golang/protobuf v1.5.1
 	github.com/google/go-cmp v0.5.6
@@ -93,10 +86,7 @@ require (
 	github.com/jmhodges/levigo v1.0.0 // indirect
 	github.com/jtacoma/uritemplates v1.0.0
 	github.com/kr/pretty v0.2.1
-	github.com/labstack/echo/v4 v4.1.16
-	github.com/labstack/gommon v0.3.0
 	github.com/lib/pq v1.10.1
-	github.com/mattn/go-isatty v0.0.12 // indirect
 	github.com/mitchellh/mapstructure v1.4.1
 	github.com/mohae/deepcopy v0.0.0-20170929034955-c48cc78d4826
 	github.com/nats-io/nats-server/v2 v2.2.2
@@ -167,7 +157,6 @@ require (
 	github.com/cespare/xxhash/v2 v2.1.1 // indirect
 	github.com/couchbase/vellum v1.0.2 // indirect
 	github.com/cpuguy83/go-md2man/v2 v2.0.0 // indirect
-	github.com/dgrijalva/jwt-go v3.2.0+incompatible // indirect
 	github.com/dgryski/go-rendezvous v0.0.0-20200823014737-9f7001d12a5f // indirect
 	github.com/dlclark/regexp2 v1.4.1-0.20201116162257-a2a8dda75c91 // indirect
 	github.com/emersion/go-sasl v0.0.0-20200509203442-7bfe0ed36a21 // indirect
@@ -190,7 +179,6 @@ require (
 	github.com/klauspost/compress v1.11.12 // indirect
 	github.com/kr/text v0.2.0 // indirect
 	github.com/magiconair/properties v1.8.1 // indirect
-	github.com/mattn/go-colorable v0.1.2 // indirect
 	github.com/mattn/go-runewidth v0.0.7 // indirect
 	github.com/matttproud/golang_protobuf_extensions v1.0.1 // indirect
 	github.com/minio/highwayhash v1.0.1 // indirect
@@ -214,8 +202,6 @@ require (
 	github.com/steveyen/gtreap v0.1.0 // indirect
 	github.com/subosito/gotenv v1.2.0 // indirect
 	github.com/tinylib/msgp v1.1.0 // indirect
-	github.com/valyala/bytebufferpool v1.0.0 // indirect
-	github.com/valyala/fasttemplate v1.0.1 // indirect
 	github.com/vmihailenco/tagparser/v2 v2.0.0 // indirect
 	github.com/willf/bitset v1.1.10 // indirect
 	go.etcd.io/bbolt v1.3.5 // indirect
