+++ pkg/util/test/assertions/should/should.go
@@ -183,10 +183,6 @@ var (
 	HaveEmptyDiff = testassertions.ShouldHaveEmptyDiff
 	// NotHaveEmptyDiff receives exactly two parameters and does an inequality check using pretty.Diff.
 	NotHaveEmptyDiff = testassertions.ShouldNotHaveEmptyDiff
-	// HaveRoute asserts that the given *echo.Echo server has a route with the given method and path.
-	HaveRoute = testassertions.ShouldHaveRoute
-	// NotHaveRoute asserts that the given *echo.Echo server does not have a route with the given method and path.
-	NotHaveRoute = testassertions.ShouldNotHaveRoute
 
 	// ResembleEvent receives exactly two events.Event and does a resemblance check.
 	ResembleEvent = testassertions.ShouldResembleEvent
