--- pkg/web/middleware/recover.go
@@ -1,47 +1,0 @@
-// Copyright © 2019 The Things Network Foundation, The Things Industries B.V.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package middleware
-
-import (
-	"fmt"
-	"os"
-	"runtime/debug"
-
-	echo "github.com/labstack/echo/v4"
-	"go.thethings.network/lorawan-stack/v3/pkg/errors"
-)
-
-// ErrHTTPRecovered is returned when a panic is caught from an HTTP handler.
-var ErrHTTPRecovered = errors.DefineInternal("http_recovered", "Internal Server Error")
-
-// Recover returns Echo middleware that recovers from panics.
-func Recover() echo.MiddlewareFunc {
-	return func(next echo.HandlerFunc) echo.HandlerFunc {
-		return func(c echo.Context) (err error) {
-			defer func() {
-				if p := recover(); p != nil {
-					fmt.Fprintln(os.Stderr, p)
-					os.Stderr.Write(debug.Stack())
-					if pErr, ok := p.(error); ok {
-						err = ErrHTTPRecovered.WithCause(pErr)
-					} else {
-						err = ErrHTTPRecovered.WithAttributes("panic", p)
-					}
-				}
-			}()
-			return next(c)
-		}
-	}
-}
