+++ pkg/web/oauthclient/cookie_auth.go
@@ -19,7 +19,7 @@ import (
 	"net/http"
 	"time"
 
-	echo "github.com/labstack/echo/v4"
+	"go.thethings.network/lorawan-stack/v3/pkg/errors"
 	"go.thethings.network/lorawan-stack/v3/pkg/web/cookie"
 )
 
@@ -43,6 +43,8 @@ type authCookie struct {
 	Expiry       time.Time
 }
 
+var errNoAuthCookie = errors.DefinePermissionDenied("no_auth_cookie", "no auth cookie")
+
 func (oc *OAuthClient) getAuthCookie(w http.ResponseWriter, r *http.Request) (authCookie, error) {
 	value := authCookie{}
 	ok, err := oc.AuthCookie().Get(w, r, &value)
@@ -50,7 +52,7 @@ func (oc *OAuthClient) getAuthCookie(w http.ResponseWriter, r *http.Request) (au
 		return authCookie{}, err
 	}
 	if !ok {
-		return authCookie{}, echo.NewHTTPError(http.StatusUnauthorized, "No auth cookie")
+		return authCookie{}, errNoAuthCookie.New()
 	}
 	return value, nil
 }
